services/llm-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ollama
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx
prometheus-client==0.19.0
structlog==23.2.0